♻️ refactor(#90): 클래스명 변경
